[FEATURE] Simulation with Akka actors
This adds a first draft of a simulator model based on Akka actors, where
messages are used for communication. The current model is time-based,
with a central timer actor that all components need to know. The
reference to the timer is given as a constructor parameter, but this
could well be replaced with using an actor ref.

The simulation currently does not end when no more steps are scheduled;
this could be improved with a timeout or an "I’m done" message sent from
actors to the scheduler.

There are currently two very simple actors that just drop a message on
the console and reschedule themselves four times, using TimerRequest
messages to the timer.