[BUGFIX] Simulation init + end not working correctly
The simulation is very sensitive to incorrect ordering of commands sent.
Therefore, we introduce a new Start() message that triggers starting the
simulation. The actors are responsible for scheduling themselves for the
first time then, as otherwise the simulation won’t be stopped.
Theoretically, they could also issue a Pass() message, that disables
them for the rest of the simulation.