[TASK] Add basic structure for RoadHopper server
This adds a custom module–mainly copied from GraphHopper–to run our own
RoadHopper server including custom modules. Currently, it replicates the
existing functionality, but the plan is to split the copied code apart
as much as possible to

a) integrate custom information for our routing application into responses
b) port senseful modularization back into GraphHopper

This is an intermediate commit to get the current state under version
control. The code should, where stated, mostly be as it was copied from
GraphHopper, especially in RoadHopperServlet.