[FEATURE] Deliver calculated road segments to client
This changes the route endpoint to deliver a list of GeoJSON constructs
based on the route calculated by RouteFactory. This makes it possible to
closely reflect the route that the vehicle will actually travel in the
simulation, and improves debugging via the map.

The current simplification threshold of 2.0 degrees is good for usage in
a city, but for Autobahn segments, sometimes too much is cut off.
Presumably this is because the optimisation uses the already averaged
segments when more segments are merged after the first two.