[FEATURE] Mark stop signs in map
This adds support for stop signs in the same way that traffic lights are
supported, with the same highlighting mechanism in the map and the same
code structure to get the information into the route export data.

The stop sign icon is licensed under Creative Commons Attribution 3.0
Unported (CC-BY 3.0), see
<https://www.iconfinder.com/icons/32553/abort_cancel_control_halt_ignore_sign_stop_stop_sign_terminate_icon#size=24>
for the source.