[FEATURE] Draw angle bisector for turns
The map now draws red lines at the intersection of two road segments,
showing the angle between the two segments. The length of the line is
proportional to the length of both segments, but the lines are only
drawn if one of the segments is short (to not get really big lines all
throughout the map).

The lines are drawn on a separate layer that can also be turned off
easily.