[TASK] Factor out road bend construction
It now brings a more realistic length calculation by ignoring the first
and last segment (which are often pretty long) and instead doubling the
second and second to last element respectively.

This method is still not perfect in detecting the bends correctly,
especially when it comes to keeping apart two consecutive bends with a
long distance between them (see this point list for an example:
point=49.006739%2C8.400904&point=49.00727%2C8.406579).