[FEATURE] Update vehicle position
The position is stored by the Journey, as it knows about the road. The
vehicle is informed about each position update with a message, currently
not including the time (but it should probably be a part of it, though
the error incurred might be negligible when doing multiple updates per
simulation second).