[BUGFIX] Simulation uses wrong order of road segments
The first segment was travelled twice, while the last one was not
travelled at all, because the stop criterion was too lax. The criterion
is now moved as a whole to the comparison with the vehicle’s current
status, as this is the only point where we can reliably detect if a
segment has been passed.

An alternative (and probably better) approach could have been to move
the distance tracking as a whole to JourneyActor, so the vehicle does
not need to keep track of the travelled distance anymore (which is not
part of its job, strictly speaking).

The travelling definitely needs a test to check if the consecutive
points are in a sensible range after each other (and no point is
travelled twice with some distance).