[FEATURE] Display simulation progress in map
This is a heavy (and quite hacky) change to the whole frontend
rendering: The server-side output is completely replaced by a custom
servlet that only includes the data required for our rendering (the
rendering code of GraphHopper cannot be called with this, so mind the
required changes there!).

The frontend rendering is really hackish and should only be seen as a
PoC. Most of it is copied from
<http://leafletplayback.theoutpost.io/examples/example_2.html>, while
the simulation.js code itself is completely new. The example’s HTML/CSS
integration requires Bootstrap, which is added temporarily, but should
be replaced with a clean slate approach instead—that is, we should roll
our own frontend as a whole and not rely on the one delivered by
GraphHopper at all. OSRM frontend v2 could be a possible base for this,
see http://project-osrm.org/osrm-frontend-v2.