[TASK] Register actors one-by-one
In the default simulation, each actor is now registered independently
with the timer. Each of these registrations is also acknowledged by the
timer, so behaviour for one-by-one and all-at-once registration is the
same—and we can use the same future for both these.

A further optimization could be to send all registrations, collect the
futures and wait for them afterwards, before returning from the
constructor. The effect should be negligible, but it might be a good
place to test this pattern.