[BUGFIX] Driver does not wait for RoadAhead results
For the driver’s interactions, we need several consecutive messages to
different actors, where the results of earlier calls are the basis of
later calls. The old structure that this commit removes does the
consecutive calls nested in the first one’s andThen(). The futures
created in there are not available to the parent method of stepUpdate(),
which waits for the futures to complete and then informs the timer that
the component is finished.

The new approach is to have a flat structure of futures. If any other
step depends on the result of one of these futures, flatMap() can be
used to glue them together. flatMap takes the result of the first future
(once it completed) and inserts it as a parameter into the method called
for the second future (the one mapped onto the first).