[TASK] Add test FSM to demo simulation
This is an attempt to test the new approach for dealing with FSMs within
the simulation. Still missing is a proper scheduling of messages for the
future; the integrated timer cannot be used for that, as it uses system
time and not simulation time. We also need to check which messages the
FSM will be able to retrieve, as it does not inherit from
SimulationActor (and probably cannot do this becaues the FSM trait
implements many methods as final).