[FEATURE] Real velocity control for driver
Until now, the driver just accelerated the vehicle without noticing the
environment. Acceleration was shut off by the vehicle when the
configured maximum speed was reached, there was no real feedback loop
between driver and vehicle.

This commit makes use of the already existing VelocityControl state
machine and extends it with proper reactions to all kinds of situations.
The general idea is to either set a target velocity or define a position
to stop at. The actor will then check with every data update how far the
defined target state is still away and will react accordingly. This
means it will adjust the acceleration until the defined velocity was
reached, or it will slow the vehicle down before approaching the stop
point. After the point has been reached, the vehicle is accelerated
again.

In this model, the driver is responsible for passing incoming
information on to the velocity control, which it does by requesting the
journey status and the road immediately ahead in its “act” step. The
driver itself does not influence the process currently after it has
initially set the target velocity (to fixed 15 m/s currently).

The vehicle will not stop directly at a defined location, but instead
about 1m before it, to account for possible inconsistencies with map
data. This also makes it easier to start again, as we don’t need to take
into account that we might already have crossed the line to the next
road segment. Accordingly, the journey checks if the vehicle is within a
1.5 m range around the end of the last segment to check if the journey
has ended (and it also checks that the vehicle is not moving anymore).

Another important change is in the vehicle’s acceleration behaviour: it
now handles positive and negative accelerations separately. A negative
acceleration (= deceleration, a braking force) will only slow the
vehicle down to 0 m/s, but not to negative speeds. This is a first small
step to a more realistic acceleration model.