[BUGFIX] Length of previous road segments tracked incorrectly
The length of previous segments was approximated by using the current
position directly after the segments were switched. The offset was
negligible for a few long segments, but if there are many short segments
(like in a roundabout), the differences might add up. Therefore, we now
use the accurate offset instead.