[FEATURE] Type-safe signals
They are kind of type safe; the type is not explicitly stored anywhere,
but we rely on Scala’s general type safety mechanisms. Unfortunately,
those only come into play at runtime in this case, as we do runtime type
conversions. So watch out to always use the correct type for one signal!