[FEATURE] Asynchronous simulation run
The simulation is now only started by the servlet, but the result is not
awaited anymore. The servlet reports back an id, that can be sent to
/roadhopper/simulationstatus. This URL should be polled until it reports
the simulation to be finished, in which case it will send the results as
done before.

The UI currently does not use this new feature, therefore simulations
with displaying the results are impossible for now.