[FEATURE] Active brake for driver
The driver can now control both the brake and gas pedal, giving them
better control over the vehicle’s movements. The brake input is the
negative parts of the PID controller’s output, limited to 0..100.
Currently, the same amplification is applied to both gas and brake pedal
input (signals alpha/beta), but this might also change. The value used
for the maximum braking force is just a guess currently.