[FEATURE] Store measurements in database
This adds a database storage for the results of measurements, supplied
as CSV files. The data must be imported during the Importer class.

Currently missing is a proper frontend implementation that allows fast
switching between measurements.

The general implementation of the measurements data might still have
problems and needs to be extended.