[TASK] Use measurement road for simulation
This adds JS code that converts the first and last point of a
measurement to regular route coordinates. The used data is the one
matched against OSM, not the original measurement positions, so expect a
small skew depending on the OSM edges used to model the road.

The GraphHopper JS code automatically does a routing based on the new
points, so a simulation can be started immediately with the points.