[BUGFIX] Remove duration from Measurement class
The additional property causes an error when serializing an object for
the database. The error also occurs when using a def instead of a lazy
val, so we remove it altogether for now.

Probably an implicit conversion to an object that has the required
method would work, or adding the method via some Scala magic.