[FEATURE] Respect road grade in simulation
The simulation now properly reads all the point’s evelation from SRTM
and calculates the grade based on this. For slow segments, the grade is
ignored, as will not contribute a lot to the actual cycle and
additionally will cause confusion if the measured data from SRTM is
inaccurate (imagine a ten meters short segment where there is a gap of
ten meters in the data—100% grade).

The grade is also used in the simulation, to calculate the climbing
resistance and lower the rolling resistance accordingly.