[FEATURE] Stop at stop signs
This commit extends the velocity estimator to look for stop signs in the
near distance. If one is encountered, the vehicle is stopped and after
one second accelerated again.

The implementation might still have its flaws, e.g. the vehicle
sometimes starts rolling backwards—preventing this would require an
addition to the brake’s input signal to always apply braking force while
the vehicle should stand still. For now, we accept the small negative
velocities, as they won’t affect the simulation too much and are also
not completely unrealistic in general.