[TASK] Lock farthest lookahead position
The new code helps preventing overshooting during a slowdown, where a
reduction of velocity leads to a reduction of the lookahead distance to
a point where the reason for the slowdown went out of sight, leading to
an increase in speed again. This led to saw teeth appearing in the
velocity diagram.